Support alternative root directory
In some cases, it's useful to be able to use configuration files
relative to a different root directory (for instance, when modifying
the network configuration inside a virtual image mounted somewhere
locally).

To use this new feature, you can invoke the wicked client with
--root-directory "/foo/bar/baz". It will pass this on to the wicked
server as an option. When accessing REST nodes below /config, it will
then use this root option to relocate all file I/O relative to this