netcf shim: actually heed the root directory
The ncf_init() function takes an argument specifying the root directory
(i.e. where to go looking for the system config files). We need to pass
this to the wicked server.

This patch adds ni_indirect_set_root() to set the root_dir on an indirect
handle, and makes sure the root directory is passed in every WICKED
request.

Signed-off-by: Olaf Kirch <okir@suse.de>