Retry failed leases in ifup processing
When wicked receives a request to reconfigure an interface, we
usually check whether we've already configured the device for the
specified addrconf mechanism. If we have, we don't touch that
service.

However, if we failed to acquire the lease, the addrconf mode
will still be set, and in that case we wouldn't retry to get
the lease. This patch adds a check for a lease state != GRANTED,
and if so, tries to re-acquire the lease.

Signed-off-by: Olaf Kirch <okir@suse.de>