netcf xml: fix two issues with representation of addrconf modes
When rendering the interface configuration as XML, we should render
an addrconf mode that is enabled but has no specific request associated
as an empty <foobar/> element.

When parsing a lease, we would attach it to the interface object using
ni_interface_set_lease. This has an undersirable side effect, however:
If the lease has a state != granted, this would disable the addrconf
mode for this interface; which is not what we want. Thus, open-code
the attachment of the lease.

Signed-off-by: Olaf Kirch <okir@suse.de>