ifconfig: when asked to configure an interface type we don't know, do not error out
Previously, we returned an error if the user tried to configure an interface
type we didn't know about, such as wireless. This patch changes this to a
debug message, and proceeds with configuring the device. This allows us to
at least take down eg a wireless interface, even if we cannot configure it
properly.

Signed-off-by: Olaf Kirch <okir@suse.de>