Changed the call signature of REST API functions
Up to now, all REST functions took an additional "path" argument,
providing the unparsed portion of the path. This was used to
access individual interfaces, which of course do not have a
static representation in the REST namespace.

The new approach is to have an optional "wildcard" child in each
node of the REST namespace. When encountering a path element that
we don't know, and there is such a "wildcard" child, we simply store
the path component in a new argv[] vector in the ni_wicked_request,
and proceed with the wildcard node.

Signed-off-by: Olaf Kirch <okir@suse.de>