ifup/down: Be less rigid with state dependencies
For instance, if  we want to configure a vlan device, we can *create*
the device as soon as the eth device is available.  And we should do
this, even is the link is not there. Lack of link on the eth device
should only prevent us from bringing up the vlan link

Signed-off-by: Olaf Kirch <okir@suse.de>