Shut up one false error message in ni_dbus_serialize_xml
When we do a GetManagedObject call, we receive a list of all interfaces
registered for the object. Not all of these may actually contain properties,
but we always tried to retrieve the "properties" type for each interface.
If there was none defined, we would complain.

We now check whether the dict returned by the server contains any
properties at all - and if it doesn't, we return quietly and do not
print any errors at all.

Signed-off-by: Olaf Kirch <okir@suse.de>