Change how we create netif objects on the client side in GetManagedObjects
We do this in two pieces. When we instantiate an object as a child of a
list object (such as Wicked/Interfaces), we automatically assign the
default list item class to the new child.

In a second step, we check if the new child has an initialize member
function, and if it has, we use that to create a local netdev object
and assign that to the proxy object.

Signed-off-by: Olaf Kirch <okir@suse.de>