Fix invalid access to free'd memory when processing the response of a DBus call
ni_dbus_object_call_variant would deserialize the message and store everything
in a (tree of) ni_dbus_variant's. However, we record the keys of a dict as just
the pointer rather than doing a strdup of the string - which is kind of bad if
ni_dbus_object_call_variant then throws aways the response message.

For this reason, introduce a __message pointer in ni_dbus_variant where we keep
a pointer to the message this variant came from. The message is dereferenced when
we throw away the variant eventually.

This is obviously a time/space tradeoff; we trade the extra pointer in the variant
for the time it takes to copy lots of little strings. Not sure what's better;
maybe at one point we need to revisit this.

Signed-off-by: Olaf Kirch <okir@suse.de>