Fix the modem event sending code
When sending a deleteDevice event for a modem, there is a race condition
because we delete the server-side dbus object after sending the event(s).
It may happen that the client receives the deleteDevice event, deletes its
local object handle, then calls GetManagedObject, which still finds the
active server-side object and returns it - appearing as a *new* object
on the client side.

This patch changes this by
	- unregistering the server-side object before sending events
	- modifying the modem event sending code to take a dbus object
	  as argument, rather than the modem device handle

Signed-off-by: Olaf Kirch <okir@suse.de>