Make use of schema information explicit in ni_dbus_{service,method}_t
Up to now, we stored the pointer to the schema information in a
"void *user_data" pointer inside method and service objects. Which
is a bit pointless, given that there's only one user of that pointer.

Make the purpose of this field explicit by using the proper
ni_xs_{method,service}_t types, and calling them "schema" rather
than "user_data".

Signed-off-by: Olaf Kirch <okir@suse.de>