network-nanny: isolate secrets set by different users
Introduce a "nanny-user", and associate a separate secret-db with
each user. In Nanny.addSecret, new secrets are added to the DB
for the caller of the method. Similarly, when updating a profile, store
the caller's uid in the profile.

Then, when we bring up an interface based on a given profile, look
up any secrets in the DB of the user who last updated the profile.

Signed-off-by: Olaf Kirch <okir@suse.de>