network-nanny: Do a better job handling device enablement
Up to now, you had to be root in order to enable a device, and
it always had to happen at runtime.

Now you can control behavior through the config file:

   <nanny>
     <allow link-type="wireless" />
     <allow class="modem" auto="true" />
     <allow device="eth0" />
   </nanny>

A device matched by such an "allow" element is now subject to
user management through network-nanny. By default, the device
still has to be enabled explicitly for nanny to monitor it;
but now any user can call "wicked nanny enable $DEVICE" on such
a device.

If the auto="true" attribute is given, then a matching device is
set up for monitoring automatically when detected.

The root user is allowed to enable or disable any device.

Signed-off-by: Olaf Kirch <okir@suse.de>