dbus: changing the signature of the generic property get_handle() function
Previously, by calling the get_handle() of function of any property belonging
to the Ethernet class, we would set the dev->ethernet handle of the underlying
netdev - no matter if we were trying to set the property (where this is the
expected behavior) or reading the property (where it is not).

This patch introduces a boolean argument "write_access", which enables the
get_handle() function to tell read from write requests.

Signed-off-by: Olaf Kirch <okir@suse.de>