When getting all properties of an object, optimize the case where get_handle() returns NULL
Quite often, we have a set of values attached to a netdev object
that is accessed through a pointer (dev->foobar).

Generic properties use the get_handle function to retrieve that
pointer, and then operate on a variable at a specific offset.

Now, if the device's pointer is not set, we could call each
property's get() function in turn, setting up a variant variable,
only to find that the get_handle() function returns NULL.

We try to optimize this case slightly by recording the address of a
failing get_handle function, and checking for this in future operations.

Signed-off-by: Olaf Kirch <okir@suse.de>