Obtain device name for installing system updates.
Since we want to migrate from writing a resolv.conf file directly to
using /sbin/netconfig, we need to grab the interface name from the lease
that is being installed.

TODO: on restoration/deletion of system information, we need to again grab
the interface name to pass to /sbin/netconfig. The tricky part is that the
lease struct in this case is gone...