API change for updating system info from lease
When a lease is removed and detached from a device, it is no longer
present in ni_netdev_t dev->leases. In order to be able to clear the
system of this lease (ie. dns/resolver settings), we need to pass the
released lease along with device name, to ni_system_update_from_lease().
We can then check it's state, and if released, forcefully add it to a specific
set of updaters to be worked on during ni_system_update_all().