Add concept of max source weight
As we add updater sources (wrappers around leases) to a particular updater,
track the max weight/importance of a source. This will be used to compute
the list of leases whose information we want to install to the system.

Make dhcp4 and dhcp6 the same weight to ensure both get installed via our
extension scripts.