Function to compute list of updater sources
When dealing with multiple leases of the same weight attached to an updater,
we may wish to install more than just one. As such, a single "best"
updater_source/lease is not enough. This function computes an array
of updater_source pointers of the highest available weight for an updater.
The system update can then act on these sources.