[system_update] generic system updater
The generic updater is meant to replace the resolver. It uses
extensions/netconfig, which relies on /sbin/netconfig, to configure
options such as DNS, NIS, NTP, etc.

Currently, the updater understands the leaseinfo (dhcpcd-test) style
data format as would be configured by the format="info" parameter in
server.xml[.in]. Addition of other formats may be needed in the future.

Please note, that this updater is not enabled as of this commit. A
future commit will replace the resolver system-updater entry in server.xml.in
with:

  <system-updater name="generic" format="info">
    <action name="backup" command="@wicked_extensionsdir@/netconfig backup"/>
    <action name="restore" command="@wicked_extensionsdir@/netconfig restore"/>
    <action name="install" command="@wicked_extensionsdir@/netconfig install"/>
    <action name="remove" command="@wicked_extensionsdir@/netconfig remove"/>
  </system-updater>