dhcp4: accept dns-domain list compatibility abuse
The dhcp dns-domain/domain-name option [15] is defined to contain
one single domain name and the dns-search option [119] a list.
Accept a domain name list in the dns-domain option from servers
sending a list for compatibility with attic dhcp clients and log
this abuse as debug. Further, fixed to not require a dns server
list to accept a domain name / search list (bnc#862130).