Add argument check to wicked xpath subcommand
Wicked xpath subcommand does not return when invoked without any
expression argument. It hangs in a read() system call.
This patch checks if an argument is given, but does not check
at this point if the expression is valid.

Also unifying the options definition and adding support to invoke help
message in two styles as the main wicked command has implemented.
   wicked xpath --help
   wicked xpath help