iflist: restore device link type
On entry into __ni_process_ifinfomsg_linkinfo(), track original device's
link.type value and restore on exit if not NI_IFTYPE_UKNOWN (ie. a new
device). There were cases when a vlan link was set to NI_IFTYPE_ETHERNET
and not set back due to missing tb[IFLA_LINKINFO] data. This caused the
vlan to be treated as an ethernet device, acquiring wake-on-lan settings
and other things.