addrconf: clean up incomplete leases on ifdown
When you bring up an interface with optional dhcp4 and dhcp6, one
of the leases may get stuck in some weird sort of limbo state on
shutdown.  In this case, the dhcp6-supplicant may return an error
NI_ERROR_ADDRCONF_NO_LEASE, but wickedd still has a record of this
incomplete lease. This patch makes sure we clean up these leftovers.

Signed-off-by: Olaf Kirch <okir@suse.de>