ipv6: wait for addresses to become usable (bnc#882259)
Added a temporary applying lease state, which is used while
applying a new granted lease to the system.

When a granted ipv6 lease arrives and the addresses are set,
we've to wait unti the kernel verifyed them for duplicates
and removes the tentative flag before we continue to apply
the rest of the lease and finally inform the client about.

When the kernel detects duplicates, the state changes either
to failed and we do not set further settings or to requesting
state for dhcpv6 leases as it automatically declines.