config: support a local.xml to override settings w/o getting in trouble with maintenance updates
If a user modifies one of the wicked xml config files, we're running
into trouble on the next maintenance update. Either the poor guy/gal is
using all modifications, or we're not applying our change.

One way around that is to have a dangling <include> element in common.xml:

  <!-- If you want to override configuration options
       in this file, consider putting them into
       local.xml rather than here.
    -->
  <include name="local.xml" optional="true"/>

We don't ship that local.xml. If it doesn't exist, we don't complain,
but if the user does create it, we parse it like any other config file.

Signed-off-by: Olaf Kirch <okir@suse.de>