ifconfig: handle link up of externally enslaved devices
In the case of standard openvswitch bridge ports, they are enslaved
by openvswitch itself to the ovs-master interface (kernel datapath
interface).
In such scenarios, we previously skipped the linkUp process, as Wicked
assumed that if a device was already enslaved, it was done via
ni_system_interface_enslave(), which generates a netlink message to
enslave and perform a link-up in bridge case, and a bonding master
manages the link state itself.

To get around this, if a device has a master, we should first check to
see if it's in link-up state. If yes, then it was enslaved by Wicked
and no furhter processing is needed. If not, then it was enslaved
externally and we need to perform the basic link-up.