nanny: Fix ni_nanny_identify_node_owner() exit condition
The function is calling it self recursively with `node->parent`, but
didn't checked if node is already a root one.

This just fix the `Segmentation fault` which was discovered, if we reach
a configuration constraint like:

```xml
<password type="string" constraint="required">
  <meta:user-input type="password" prompt="please enter WPA EAP password"/>
</password>
```
and the password wasn't set in the current configuration.

But it doesn't fix the prompt for password!