wpa: Use new DBus API from fi.w1.wpa_supplicant1
Change the DBus interface of wpa_supplicant to v1 version
(see: https://w1.fi/wpa_supplicant/devel/dbus.html).

Following interfaces are implemented/used:
    * fi.w1.wpa_supplicant1
    * fi.w1.wpa_supplicant1.Interface
    * fi.w1.wpa_supplicant1.BSS
    * fi.w1.wpa_supplicant1.Network

The idea is, that wpa-supplicant.[ch] is considered as helper to
communicate with the wpa_supplicant service, but do _not_ have a 1:1 relation
to wireless.c, because we would like to use it later with 802.1X on
ethernet as well.

A user can register various callbacks to get status and property updates
like:
  * on_network_added
  * on_network_selected
  * on_state_change
  * on_properties_changed

As we do not write every property of a fi.w1.wpa_supplicant1.Network we
can not "update" such a DBus object, we need to delete and recreate
them.
A network is created on `DeviceChanged` event, but it is `disabled`. On
wireless_connect() all networks get enabled.