dbus: Fix common dbus string_array handling
We missed the cleanup of the old value. This lead into
orphan entries in our local datastructure.

It was detected when debugging BSS info on wireless.
If I kept the same SSID/BSSID, but changed pairwise or key management
settings, the old values where still displayed. Because the dbus
object just get updated.