wireless: Set key-management default based on capabilities
We use the passphrase to detect a PSK configuration and the
eap-method for EAP.
Once a config is detected and the key-management wasn't set
explicit, we use defaults for PSK or EAP. The default is all
known suitable key-managment methods, but only those who
are mentioned in capabilities.

PMF will be enabled (optional), once it isn't set via config
and a "newer" key-management method is used.