firewall-ext: skip firewallDown action without previous firewallUp
Reference: bsc#1201053

If a interface configuration contains FIREWALL=no, wicked should not
touch the firewall configuration on firewallUp _and_ firewallDown.

The firewall schema contains a `skip-unless-present="true"` for the
method firewallUp. This avoid the calling of the firewallUp DBus method
on the specific state, as long as the configuration doesn't contain the
<firewall> node.

On firewallDown, we do not have the configuration, thus we cannot use
`skip-unless-present="true"` here. And the firewallDown state calls
the firewallDown DBus method regardless of FIREWALL configuration value.
To avoid unexpected firewall changes, the firewall extension handle
firewallDown events only, if a corresponding firewallUp event was
invoked before.