ethtool: remove link mode mask nwords calculations
Remove the nwords pre-calculation from the link mode constants
in the linux/ethtool.h header. We have to query it and use the
nwords value of the kernel for subsequent calls and already do.
Further, don't limit the list of advertise link modes on dbus
to mode count in a potentially outdated header file, but just
apply to the bitfield; we filter driver supported modes later.