__ni_rtnl_parse_newaddr: Use peer-to-peer condition from comment instead of relying on interface's p2p flag
An ethernet interface is not a peer-to-peer interface. However, we may bind peer to peer addresses on it, which won't get removed with the current code (ni_nl_talk() in __ni_rtnl_send_deladdr() returns an error because ap->peer_addr stays empty for that p2p-address).

Instead, use the peer-to-peer condition from the comment below to check whether the interface address is P2P.