fsm: removed modifications to applied configs
Generation of the configs on the fly in the fsm is too
late and causes side effects, that we delete a policy,
but the fsm may still generate config via dependencies.
Instead, we generate the missing configs/policies just
after parsing configuration files in migration step.
Also don't remove master reference from applied config
when clearing/resetting the interface hierarchy.