appconfig: increase arp (verify|notify) retry value (bsc#1218668)
When the kernel report ENOBUFS while sending the arp package, we
do not want to treat this as a fatal error. But we also don't want
to accept ENOBUFS forever. The previous retry value of 3 wasn't
sufficient and was triggered on bond interfaces with 802.3ad mode
to often.
With a default of 10 we give the driver a minimum amount of 6s
to accept packages.