client: fix to set and use origin in loaded config
Request to update config origin from it's source while reading
to load the config with this origin to worker and send to nanny
on ifup/ifreload or using it in ifstatus/ifcheck.
Removed the "raw" parameter from the load function completely;
it's used to clear the origin and other attributes.